+++ circle-01/Libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: jseo <jseo@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/23 19:33:43 by jseo              #+#    #+#             */
-/*   Updated: 2020/12/25 18:49:19 by jseo             ###   ########.fr       */
+/*   Updated: 2020/12/28 22:45:44 by jseo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
-	while (++i < (int)n)
+	while (++i < n)
 		if (((unsigned char *)s)[i] == (unsigned char)c)
 			return ((void *)s + i);
 	return (NULL);
