+++ circle-01/Libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: jseo <jseo@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/23 19:44:12 by jseo              #+#    #+#             */
-/*   Updated: 2020/12/25 18:49:25 by jseo             ###   ########.fr       */
+/*   Updated: 2020/12/28 22:46:00 by jseo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
-	while (++i < (int)n)
+	while (++i < n)
 		if (((unsigned char *)s1)[i] != ((unsigned char *)s2)[i])
 			return (((unsigned char *)s1)[i] - ((unsigned char *)s2)[i]);
 	return (0);
